Extract useSkipAnalysis composable from EditorApp.vue
Move skip analysis logic (~175 lines) into dedicated composable:
- isConstant, getConstantValue pure helpers
- analyzeSkipSourceCondition, analyzeSkipDestCondition
- isRowSkipped (depends on mappingDocument ref)

Zero functional changes - pure refactoring.
